Fix syntax error
You may be seeing a bunch of very trite commits due to the awful
nature of this workflow. I have one machine with no push access to test
on, and another to do development on with push access. This sucks, but
it explains why every time I forget to write self.method() I have to push stuff back and forth to remain in sync.